Create: Crafted Beginning v0.0.9 更新日志 ====================
====================
新增内容
■ 注气室滤网
可以安装在注气室底部，从而让注气室实现鼓风机类似的批量加工行为。
需要先用鼓风机的加工方式处理后才能记录对应的加工类型。

修改内容
■ 气密壳装管道
现在有更清晰的可视化显示。

■ 旋风人蓄风室
现在实时根据恢复的饥饿值、饱和度和提供的药水效果计算食物可以提供的时长。
现在自动支持其他模组添加的食物。
略微调整了食物可以通过的时长算法，并允许配方中返还一个物品。

■ 创造物品栏
现在根据"气密系列"、"光应力系列"、"终末系列"、"装饰"和"气罐"进行分类。

■ 注气室
现在允许进行批量注气，单次可加工最多一组物品。

■ 气体工厂仪表
修改合成配方，现在材料中还额外需要紫晶体板。

■ 气密板
修改序列组装配方，调换了第一个材料（铁板）和第二个材料（粘液球或者强力胶）的顺序。

■ 气球
重新调整工具栏显示，不再用耐久条显示气体含量与种类。

====================
Bug修复
■ 修复在安装iris时黑曜石砖块N、黑曜石砖块S、哭泣黑曜石砖块N和哭泣黑曜石砖块S丢失材质的问题。
■ 修复特斯拉涡轮机和气密引擎的护目镜提示框里在其他字体或英文语言环境下不对齐的问题。
■ 修复注气序列组装第一个材料显示出错的问题。
■ 修复注气室动画卡顿的问题。
■ 修复气密引擎动画卡顿的问题。
■ 修复在未安装JEI时右键打开仓库管理员的GUI会报错闪退的问题。
■ 修复气体过滤器的最大堆叠数被设置为1的问题。
■ 修复在安装流体包裹且进行JEI转单时会丢失气体材料的问题。
■ 修复气球显示气体含量时冒号的颜色也被渲染成金色的问题。

+++ src/main/java/net/ty/createcraftedbeginning/content/airtights/airtightengine/airtightassemblydriver/AirtightAssemblyDriverStructureManager.java
@@ -38,17 +38,17 @@ public class AirtightAssemblyDriverStructureManager {
         long width = AirtightTankBlockEntity.getConfiguredMaxWidth();
         long length = AirtightTankBlockEntity.getConfiguredMaxLength();
         long surfaceArea = 2L * (width * width + 2L * width * length);
-        return (int) Math.clamp(surfaceArea, 0, Integer.MAX_VALUE);
+        return Math.clamp(surfaceArea, 0, Integer.MAX_VALUE);
     }
 
     private static int getMaxAttachedChambers() {
         long width = AirtightTankBlockEntity.getConfiguredMaxWidth();
         long length = AirtightTankBlockEntity.getConfiguredMaxLength();
-        return (int) Math.clamp(width * Math.max(width, length), 0, Integer.MAX_VALUE);
+        return Math.clamp(width * Math.max(width, length), 0, Integer.MAX_VALUE);
     }
 
     private static int getMaxAttachedWindChargingLevel() {
-        return (int) Math.clamp((long) getMaxAttachedChambers() * MAX_LEVEL, 0, Integer.MAX_VALUE);
+        return Math.clamp((long) getMaxAttachedChambers() * MAX_LEVEL, 0, Integer.MAX_VALUE);
     }
 
     private final AirtightAssemblyDriverCore driverCore;
